requirements.txt
streamlit
pandas
scikit-learn
sentence-transformers
faiss-cpu
numpy
